+++ linux/net/rina/Makefile
@@ -45,6 +45,7 @@ rina-default-plugin-y :=			\
 	rmt-ps-default.o			\
 	dtp-ps-common.o				\
 	dtp-ps-default.o			\
+	dtcp-ps-common.o 		        \
 	dtcp-ps-default.o                       \
 	pft-ps-default.o
