+++ linux/net/rina/rnl-utils.h
@@ -117,7 +117,6 @@ enum dtcp_fctrl_config_attrs_lists {
         DFCC_ATTR_RBYTES_PER_THRES,
         DFCC_ATTR_RBUFFER_THRES,
         DFCC_ATTR_CLOSED_WINDOW_POLICY,
-        DFCC_ATTR_FLOW_CTRL_OVERRUN_POLICY,
         DFCC_ATTR_RECON_FLOW_CTRL_POLICY,
         DFCC_ATTR_RCVING_FLOW_CTRL_POLICY,
         __DFCC_ATTR_MAX,
